// utils/types.ts
export type PreviewImage = {
  id: string | null;
  url: string;
  backgroundColor?: string;
  color?: string;
  isLoading?: boolean;
};

export type Option = {
  label: string;
  value: string;
  isSelected?: boolean;
};

// Schema

type OrderDetails = Record<
  string,
  {
    product: Pick<Product, "title" | "price" | "quantity"> & { image: string };
    quantity: number;
  }
>;
export interface Order extends Omit<Transactions, "details" | "address"> {
  details: OrderDetails[];
  address: Address;
  user: User;
}

export interface ProductI extends Omit<Product, "images" | "slideColors"> {
  images: UploadResponse[];
  category: Record<"name" | "id", string>[];
  slideColors: Record<"color" | "backgroundColor", string>[];
  sizes: string[]; // NEW: Sizes
  keyFeatures: string[]; // NEW: Key Features
  careInstructions: string; // NEW: Care Instructions
  specifications: { label: string; value: string }[]; // NEW: Specifications

}
// ...
export type StateUpdateType = "add" | "update" | "delete";